requirements.txt
Flask==3.1.2
Flask_Cors==4.0.0
python-dotenv==1.2.1
Requests==2.32.5
supabase==2.24.0
gunicorn==21.2.0